Initial commit from a long, long time
Finally I return now to the Python Challenge. The first time I knew this
very tricky website was in 2014, first year of the college. I was
using only Python to solve the problems (which makes a lot of sense) but
I got stuck on the problem 3 for years because I just not saw the
EXACTLY word stamped on my face. Sad. This changes everything!

Well...

For now, to mix-up my rage accumulated from the years, I'm solving
all the problems using just Common Lisp. Let's all cry together!

Solutions from the problems:

* level-0
* level-1
* level-2
* level-3
* level-4